Champion Orpheus

+++ Core/src/main/java/util/Settings.java
@@ -11,10 +11,10 @@ public class Settings {
 	public static final int FPS = 30;
     
 	// seconds to frames
-	public static int seconds(int sec){
-		return sec * FPS;
+	public static int seconds(double sec){
+		return (int) (sec * FPS);
 	}
-	public static int framesToSeconds(int frames){
-		return frames / FPS;
+	public static double framesToSeconds(int frames){
+		return ((double)frames) / FPS;
 	}
 }
